Ensure that azure storage is loaded correctly
In the `azure-storage-blob` gem, that is specifically required for active storage in the README,
there are no `azure/storage` file. The `azure/storage` file was only present on the older, and now deprecated, `azure-storage` gem.

So the old require will naturally give a load error, and the old class and method references to azure storage has been changed.

At the moment a fix to this problem is underway - but currently not implemented in any official release (as of now, the latest is 6.0.3.1),
but this solution will work just fine with 5.2.4.2.